Updated README.md YouCompleteMe Section
In the "Install YouCompleteMe" section there is an implicit assumption that the user is still in the ~/.vim/bundle directory. If the user does not follow the guide exactly as outlined and changes directories between these sections, or if the user elects not to install syntax highlighting, this assumption could lead to errors for true beginners.